apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
pydantic-settings==2.8.1
httpx==0.28.1
apscheduler==3.11.0
python-dotenv==1.0.1
openai==1.68.2
agno==2.0.9
elevenlabs==1.58.1
PyJWT[crypto]==2.10.1
python-multipart==0.0.20
pypdf==5.4.0
